Moved signup and login functions out from main.js etc..
