+++ admin.html
@@ -5,12 +5,21 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Admin</title>
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.15.4/css/all.css" integrity="sha384-DyZ88mC6Up2uqS4h/KRgHuoeGwBcD4Ng9SiP4dIRy0EXTlnuz47vAwmeGwVChigm" crossorigin="anonymous">
+    <link rel="stylesheet" href="./css/navbar.css">
     <link rel="stylesheet" href="./css/adminPage.css">
 
-    <script type="module" src="./js/uploadProduct.js"></script>
-    <script type="module" src="./js/adminpage.js"></script>
+    
+    <script defer type="module" src="./components/navbar.js"></script>   
+    <script defer type="module" src="./js/uploadProduct.js"></script>
+    <script defer type="module" src="./js/adminpage.js"></script>
+
 </head>
 <body>
+
+     <!-- navbar -->
+     <div id="navbar" class="zindexThree"></div>
+
 
     <div class="wrapper justifyCenter">
