+++ api/classes/createInstanceFunctions.php
@@ -38,10 +38,10 @@ function createProductInCategory($productId, $categoryId) {
 include_once("../classes/userClass.php");
 
 function createUser($ID, $UserEmail, $UserPassword, $UserRegisterDate, $UserFirstName, $UserLastName,
-$Newsletter, $TermsConditions, $UserIsAdmin, $AdminRequest,  $UserCountry, $UserCity,  $UserStreet, $UserZipCode) {
+$TermsConditions, $UserIsAdmin, $AdminRequest,  $UserCountry, $UserCity,  $UserStreet, $UserZipCode) {
     
     return new User((int)$ID, $UserEmail, $UserPassword, (int)$UserRegisterDate, $UserFirstName, $UserLastName,
-    (int)$Newsletter, (int)$TermsConditions, (int)$UserIsAdmin, (int)$AdminRequest,  $UserCountry, $UserCity,  $UserStreet, $UserZipCode);
+   (int)$TermsConditions, (int)$UserIsAdmin, (int)$AdminRequest,  $UserCountry, $UserCity,  $UserStreet, $UserZipCode);
 }
