+++ api/controllers/authAdmin.php
@@ -11,8 +11,7 @@ if(isset($_SERVER['REQUEST_METHOD'])) {
              $userObj = unserialize($_SESSION["loggedInAdmin"]);
              echo json_encode($userObj);
             
-             
-            /*  echo json_encode(true); */
+            
      } else {
          echo json_encode(false);
      }
