+++ product.html
@@ -18,7 +18,6 @@
 
 
     <!-- JS Main -->
-    <script defer type="module" src="./js/main.js"></script> 
     <script defer type="module" src="./components/footer.js"></script>
     <!-- JS common -->
     <script defer type="module" src="./components/navbar.js"></script>
